requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
redis
httpx
python-multipart
pillow
opencv-python-headless
numpy
python-dotenv
aiofiles
jinja2